Migrate CLI from clap v2 to clap v4 builder API
- Replace clap v2 builder API with clap v4 equivalents throughout
- Add clap_complete v4 for shell completion generation in build.rs
- Convert boolean flags to ArgAction::SetTrue, verbosity to ArgAction::Count
- Replace validators with value_parser for typed argument extraction
- Update all ArgMatches extraction: value_of→get_one, is_present→get_flag,
  occurrences_of→get_count, values_of→get_many
- Remove lifetime parameters from ArgMatches<'_>
- Update logging verbosity type from u64 to u8
- Update lit test expectations for clap v4 help format

+++ src/app/logging.rs
@@ -16,14 +16,14 @@ use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::EnvFilter;
 
 pub struct Logging {
-    verbosity: u64,
+    verbosity: u8,
     rust_log: Option<OsString>,
     color: bool,
     debug: bool,
 }
 
 impl Logging {
-    pub fn new(verbosity: u64, rust_log: Option<OsString>, color: bool, debug: bool) -> Logging {
+    pub fn new(verbosity: u8, rust_log: Option<OsString>, color: bool, debug: bool) -> Logging {
         Logging {
             verbosity,
             rust_log,
@@ -32,7 +32,7 @@ impl Logging {
         }
     }
 
-    fn log_level(verbosity: u64) -> LevelFilter {
+    fn log_level(verbosity: u8) -> LevelFilter {
         match verbosity {
             0 => LevelFilter::WARN,
             1 => LevelFilter::INFO,
